Update list of announcements component

+++ client/src/app/announcements/announcement-comments/announcement-comments.component.ts
@@ -2,6 +2,7 @@ import { Component, Input, OnInit, ViewChild } from '@angular/core';
 import { NgForm } from '@angular/forms';
 import { Announcement } from 'src/app/models/announcement';
 import { CommentOfAnnouncement } from 'src/app/models/commentOfAnnouncement';
+import { User } from 'src/app/models/user';
 import { AccountService } from 'src/app/services/account.service';
 import { CommentService } from 'src/app/services/comment.service';
 
@@ -17,11 +18,15 @@ export class AnnouncementCommentsComponent implements OnInit {
   model: any = {};
   loading = false;
   InputText = EnumInputText;
+  user: User;
 
   constructor(private commentService: CommentService, public accountService: AccountService) { }
 
   ngOnInit(): void {
     this.getComments();
+    this.accountService.currentUser$.subscribe(user => {
+      this.user = user;
+    })
   }
 
   sendComment(){
@@ -36,15 +41,17 @@ export class AnnouncementCommentsComponent implements OnInit {
   getComments(){
     this.commentService.getComments(this.announcement.id).subscribe(response =>{
       this.comments = response;
+      console.log(this.comments)
     })
   }
 
   deleteComment(comment: CommentOfAnnouncement){
-    this.commentService.deleteComment(comment.id).subscribe(() =>{
-      this.comments = this.comments.filter(item => item.id !== comment.id);
-    })
+    if(confirm("Czy na pewno chcesz usunąć ten komentarz?")){
+      this.commentService.deleteComment(comment.id).subscribe(() =>{
+        this.comments = this.comments.filter(item => item.id !== comment.id);
+      })
+    }
   }
-
 }
 
 export enum EnumInputText {
